Version 1
This is the first version of my "Guess my number" for the programming assignment. In this version, I added the introduction(How the code should present itself when the game starts), the if/else statements to direct the user in whichever path he chose, and a setup for random variables.