Version 2
This is the second version of my "Guess my number" for the programming assignment. In this version, I completely fleshed out all of the 1 to 100 random choice guess route(This includes "too high", "too low" and "out of bounds"), with the counting feature that counts how many tries it takes you to get the answer. Unfortunately, I couldn't get the "play again" feature to work, and I will fix it in the later version