Version 3
This is the third version of the Guess my number. In this version, I completed the user parameter(The user chooses from what number to the other as parameters for a number to be randomly chosen from) route. It was quite similar to the other route, so copy-pasting and modifying it was good enough. The play again feature is still not fixed, and it is the last thing needed to complete this game.